% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_gradient.R
\name{compute_gradient}
\alias{compute_gradient}
\title{Computes Gradient}
\usage{
compute_gradient(a, epsilon, beta, x, y, KO_mod_mat)
}
\arguments{
\item{a}{A number in \code{[0, 1]} that represents the probability of observing a KO given
that one of the modules it is included in is present.}

\item{epsilon}{A number in \code{[0, 1]} that represents the probability of observing a
KO given that none of the modules it is included in are present.}

\item{beta}{A matrix with \code{p+1} rows and \code{m} columns. The first row are
intercepts for each module, and the element in row \code{i} and column \code{k} represents
the coefficient for the covariate \code{x_{i+1}} and module \code{k}.}

\item{x}{A design matrix, where the first column contains \code{1}'s and the next \code{p}
columns include the \code{p} covariates measured on the \code{n} samples.}

\item{y}{The outcome matrix, with \code{n} rows and \code{r} columns where the element in
row \code{i} and column \code{j} is \code{1} if KO \code{j} was observed in sample \code{i}
and \code{0} otherwise.}

\item{KO_mod_mat}{A matrix with \code{r} rows and \code{m} columns that encodes module
definitions, where a \code{1} for row \code{j} and column \code{k} means that KO \code{j}
is in the definition for module \code{k}.}
}
\value{
The gradient of negative log likelihood for a certain set of parameters and data.
}
\description{
The function computes the gradient of the negative log likelihood based on the input
data and parameter values.
}
\examples{
dat <- generate_basic_example()
beta <- matrix(rnorm(4), nrow = 2)
compute_gradient(0.8, 0.1, beta, dat$x, dat$y, dat$KO_mod_mat)

}
