% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_log_lik.R
\name{compute_log_lik}
\alias{compute_log_lik}
\title{Computes Negative Log Likelihood}
\usage{
compute_log_lik(a, epsilon, beta, x, y, KO_mod_mat, use_expit = FALSE)
}
\arguments{
\item{a}{A number in \code{[0, 1]} that represents the probability of observing a KO given
that one of the modules it is included in is present.}

\item{epsilon}{A number in \code{[0, 1]} that represents the probability of observing a
KO given that none of the modules it is included in are present.}

\item{beta}{A matrix with \code{p+1} rows and \code{m} columns. The first row are
intercepts for each module, and the element in row \code{i} and column \code{k} represents
the coefficient for the covariate \code{x_{i+1}} and module \code{k}.}

\item{x}{A design matrix, where the first column contains \code{1}'s and the next \code{p}
columns include the \code{p} covariates measured on the \code{n} samples.}

\item{y}{The outcome matrix, with \code{n} rows and \code{r} columns where the element in
row \code{i} and column \code{j} is \code{1} if KO \code{j} was observed in sample \code{i}
and \code{0} otherwise.}

\item{KO_mod_mat}{A matrix with \code{r} rows and \code{m} columns that encodes module
definitions, where a \code{1} for row \code{j} and column \code{k} means that KO \code{j}
is in the definition for module \code{k}.}

\item{use_expit}{If \code{TRUE}, then replace \code{a} and \code{epsilon} with \code{expit(a)}
and \code{expit(epsilons)}. Set to \code{FALSE} by default.}
}
\value{
The negative log likelihood of the parameters given the data.
}
\description{
This function computes the negative log likelihood for the parameters given the data.
}
\examples{
dat <- generate_basic_example()
beta <- matrix(rnorm(4), nrow = 2)
compute_log_lik(0.8, 0.1, beta, dat$x, dat$y, dat$KO_mod_mat)

}
