% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expit.R
\name{expit}
\alias{expit}
\title{Expit}
\usage{
expit(x)
}
\arguments{
\item{x}{A real number.}
}
\value{
A number representing the expit of \code{x}. This is equal to
\code{1/(1+e^(-x))}.
}
\description{
Computes the expit of a value
}
\examples{
expit(1)
expit(1000)
expit(-1000)

}
