% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_basic_example.R
\name{generate_basic_example}
\alias{generate_basic_example}
\title{Generate Basic Example}
\usage{
generate_basic_example()
}
\value{
A list including \code{x}, \code{y}, and \code{KO_mod_mat}.
}
\description{
Generates data to test that functions are working as expected
}
\examples{
generate_basic_example()

}
