% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_lambda_space.R
\name{generate_lambda_space}
\alias{generate_lambda_space}
\title{Generate lambda space}
\usage{
generate_lambda_space(m)
}
\arguments{
\item{m}{The number of modules in the dataset.}
}
\value{
A matrix with \code{2^m} rows and \code{m} columns. Each row represents one
point in the sample space for \code{lambda}. Each column contains a \code{1} for
modules that are present and a \code{0} for modules that are absent for that point in
the sample space.
}
\description{
This function generates the sample space for the vector lambda.
}
\examples{
generate_lambda_space(4)

}
