% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_lambda_space.R
\name{transform_lambda_space}
\alias{transform_lambda_space}
\title{Transform lambda space}
\usage{
transform_lambda_space(m, r, sample_space, KO_mod_mat)
}
\arguments{
\item{m}{The number of modules in the dataset.}

\item{r}{The number of KOs in the dataset.}

\item{sample_space}{The sample space for the lambda vector.}

\item{KO_mod_mat}{A matrix with \code{r} rows and \code{m} columns that encodes module
definitions, where a \code{1} for row \code{j} and column \code{k} means that KO \code{j}
is in the definition for module \code{k}.}
}
\value{
A matrix with \code{2^m} rows and \code{r} columns. Each row represents one
point in the sample space for \code{lambda}. Each column contains a \code{1} for
KOs for which one of the modules they are included in is present and a \code{0} for KOs
for which none of the modules they are included in are present.
}
\description{
This function takes the sample space for the vector lambda and transforms it into the
sample space for the maximum lambda values over all indices in the set \code{D_j} for
each KO \code{j}
}
\examples{
sample_space <- generate_lambda_space(2)
KO_mod_mat <- matrix(c(1, 1, 1, 0, 1, 0, 0, 1, 0, 1), nrow = 5, byrow = TRUE)
transform_lambda_space(2, 5, sample_space, KO_mod_mat)

}
